+++ src/main/kotlin/no/nav/familie/pdf/pdf/PdfService.kt
@@ -3,22 +3,8 @@ package no.nav.familie.pdf.pdf
 import com.itextpdf.io.source.ByteArrayOutputStream
 import no.nav.familie.pdf.pdf.PdfUtils.lagDokument
 import no.nav.familie.pdf.pdf.PdfUtils.lagPdfADocument
-import no.nav.familie.pdf.pdf.PdfValidator.validerPdf
-import no.nav.familie.pdf.pdf.domain.PdfMedStandarder
-import no.nav.familie.pdf.pdf.domain.PdfStandard
-import no.nav.familie.pdf.pdf.domain.Standard
 
 class PdfService {
-    fun opprettPdfMedStandarder(feltMap: Map<String, Any>): PdfMedStandarder {
-        val pdf = opprettPdf(feltMap)
-        val pdfMedStandarder =
-            PdfMedStandarder(
-                pdf,
-                createStandarder(pdf),
-            )
-        return pdfMedStandarder
-    }
-
     fun opprettPdf(feltMap: Map<String, Any>): ByteArray {
         feltMap.values.forEach { value ->
             requireNotNull(value) { "feltMap sitt label eller verdiliste er tom." }
@@ -28,13 +14,5 @@ class PdfService {
         lagDokument(pdfADokument, feltMap)
 
         return byteArrayOutputStream.toByteArray()
-    }
-
-    private fun createStandarder(pdf: ByteArray): Map<PdfStandard, Standard> {
-        val standarder =
-            PdfStandard.values().associateWith {
-                validerPdf(pdf, it.standard) // Use the `standard` string property here
-            }
-        return standarder
     }
 }
